Move network package to new file
